removed home page and modified principal view
